fix: drop data
